Add a statement in the very start of the course
This commit adds a very motivational and encouraging statement in the start of the course which will surely make all the course takers to finish the course.

This commit is not made to just learn how we can contribute to projects. Trust Me :)

Actually it's made for that exact same reason. :) :)